chatbot/requirements.txt
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-deepseek
chromadb
fastapi
uvicorn
python-docx
pandas
openpyxl
rank_bm25
python-dotenv
pydantic
tiktoken
pymupdf
langchain-chroma
faiss-cpu
